bora-api/requirements.txt
# Bora API — Python Dependencies

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Python 3.9 compatibility for Pydantic v2
eval_type_backport>=0.2.0

# AI
anthropic==0.34.0
openai==1.51.0

# Database
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
pgvector==0.3.2

# Embeddings
sentence-transformers==3.0.0

# Export
cairosvg==2.7.1
lxml==5.2.2

# Storage
boto3==1.34.0
supabase>=2.0.0
requests>=2.31.0

# Cache
redis==5.0.8

# Task queue
celery==5.4.0

# Auth
pyjwt==2.9.0
cryptography==43.0.0

# Validation
pydantic==2.9.0
pydantic-settings==2.4.0

# Dev/Test
pytest==8.3.0
pytest-asyncio==0.23.8
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
